+++ src/boards/wiscore_rak3401_rak13302_w25q16/pinconfig.c
@@ -1,0 +1,1 @@
+#include "../wiscore_rak3401/pinconfig.c"