+++ src/boards/wiscore_rak4631_w25q16/pinconfig.c
@@ -1,0 +1,1 @@
+#include "../wiscore_rak4631_board/pinconfig.c"